+++ src/main/java/net/sf/maventaglib/TagreferenceRenderer.java
@@ -38,7 +38,6 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.maven.doxia.module.xhtml.XhtmlParser;
 import org.apache.maven.doxia.parser.ParseException;
 import org.apache.maven.doxia.sink.Sink;
-import org.apache.maven.doxia.sink.impl.SinkEventAttributeSet;
 import org.apache.maven.plugin.logging.Log;
 import org.apache.maven.reporting.AbstractMavenReportRenderer;
 
@@ -250,9 +249,7 @@ public class TagreferenceRenderer extends AbstractMavenReportRenderer
         if (StringUtils.isNotEmpty(tag.getExample()))
         {
             startSection(Messages.getString("Tagreference.example")); //$NON-NLS-1$
-            sink.verbatim(SinkEventAttributeSet.BOXED);
-            sink.text(tag.getExample());
-            sink.verbatim_();
+            verbatimText(tag.getExample());
             endSection();
         }
 
@@ -339,9 +336,7 @@ public class TagreferenceRenderer extends AbstractMavenReportRenderer
         if (StringUtils.isNotEmpty(tag.getExample()))
         {
             startSection(Messages.getString("Tagreference.example")); //$NON-NLS-1$
-            sink.verbatim(SinkEventAttributeSet.BOXED);
-            sink.text(tag.getExample());
-            sink.verbatim_();
+            verbatimText(tag.getExample());
             endSection();
         }
 
@@ -526,9 +521,7 @@ public class TagreferenceRenderer extends AbstractMavenReportRenderer
         if (StringUtils.isNotEmpty(tag.getExample()))
         {
             startSection(Messages.getString("Tagreference.example")); //$NON-NLS-1$
-            sink.verbatim(SinkEventAttributeSet.BOXED);
-            sink.text(tag.getExample());
-            sink.verbatim_();
+            verbatimText(tag.getExample());
             endSection();
         }